Fix OAuth2 Expectations merge being in-place and then returning a copy
.Merge() used to work on the receiver, silently changing the receiver and afterwards returning a copy.
This commit changes the method to create a copy and work on the copy.
Most uses of the function are only called once, where the bug isn't triggered.
However, this prevented the JWT issuer with dynamic Issuer URLs from working correctly
because only the first issuer would be taken for eternity.